Move GIMP logic to main module
