+++ src/stx/headers/stxread.h
@@ -4,5 +4,6 @@
 #include "image.h"
 #include "value.h"
 
-StxResult<StxImage> read(std::istream &file);
-StxResult<gint32> to_image(const StxImage &data);
+namespace stx {
+  StxResult<StxImage> read(std::istream &file);
+}
