+++ src/stx/stxread.cpp
@@ -16,7 +16,7 @@ static StxError parse_geometry(const char *buffer, Geometry &geom) {
   return StxError::SUCCESS;
 }
 
-StxResult<StxImage> read(stream &file) {
+StxResult<StxImage> stx::read(stream &file) {
   using Result = StxResult<StxImage>;
 
   char magic[STX_MAGIC_SIZE];
@@ -84,59 +84,4 @@ StxResult<StxImage> read(stream &file) {
     return ERR(StxError::EARLY_EOF);
 
   return OK(data);
-}
-
-StxResult<gint32> to_image(const StxImage &data) {
-  using Result = StxResult<gint32>;
-
-  gint32 image_id = gimp_image_new(
-    data.geometry.width,
-    data.geometry.height,
-    GIMP_RGB
-  );
-  //gimp_image_set_filename(image_id, filename);
-
-  gint32 layer_id = gimp_layer_new(
-    image_id,
-    "Texture",
-    data.geometry.width, data.geometry.height,
-    GIMP_RGBA_IMAGE,
-    100.0,
-    GIMP_NORMAL_MODE
-  );
-  gboolean success = gimp_image_insert_layer(image_id, layer_id, 0, 0);
-  
-  if (!success) {
-    gimp_image_delete(image_id);
-    return ERR(StxError::GIMP_ERROR);
-  }
-
-  GimpDrawable *drawable = gimp_drawable_get(layer_id);
-
-  GimpPixelRgn rgn_out;
-  gimp_pixel_rgn_init(
-    &rgn_out, drawable,
-    0, 0, data.geometry.width, data.geometry.height,
-    TRUE, TRUE
-  );
-
-  gimp_pixel_rgn_set_rect(
-    &rgn_out, data.image_data,
-    0, 0,
-    data.geometry.width, data.geometry.height
-  );
-
-  delete[] data.image_data;
-
-  gimp_drawable_flush(drawable);
-  gimp_drawable_merge_shadow(drawable->drawable_id, TRUE);
-  gimp_drawable_update(
-    drawable->drawable_id,
-    0, 0,
-    data.geometry.width, data.geometry.height
-  );
-
-  gimp_drawable_detach(drawable);
-
-  return OK(image_id);
 }
