+++ src/stx/stxwrite.cpp
@@ -7,15 +7,6 @@
 
 using stream = std::ofstream;
 
-StxParams default_params() {
-  StxParams params;
-  params.scale_x = 100;
-  params.scale_y = 100;
-  params.e6_write = FALSE;
-  params.magical_number = 4;
-  return params;
-}
-
 #define CONVERT_SCALE(x) 25600 / x
 
 static void build_geometry(
@@ -38,7 +29,7 @@ static void build_geometry(
 
 #define DEFAULT_NUMBER '\x04'
 
-StxResult<std::monostate> write(
+StxResult<std::monostate> stx::write(
   const StxImage &img,
   std::ostream &file
 ) {
@@ -77,52 +68,4 @@ StxResult<std::monostate> write(
 
   std::monostate unit;
   return OK(unit);
-}
-
-StxResult<std::monostate> write(
-  const StxParams &params,
-  gint32 drawable_id,
-  std::ostream &file
-) {
-  using Result = StxResult<std::monostate>;
-
-  StxImage img;
-
-  guint16 width = gimp_drawable_width(drawable_id);
-  guint16 height = gimp_drawable_height(drawable_id);
-
-  img.geometry = {
-    width,
-    height,
-    params.scale_x,
-    params.scale_y
-  };
-
-  const bool has_alpha = gimp_drawable_has_alpha(drawable_id);
-  const size_t offset = has_alpha ? 4 : 3;
-
-  GimpPixelRgn rgn_in;
-  GimpDrawable *drawable = gimp_drawable_get(drawable_id);
-
-  gimp_pixel_rgn_init(&rgn_in, drawable, 0, 0, width, height, FALSE, FALSE);
-  size_t pixels = width * height;
-  guchar* data = new guchar[pixels * offset];
-  gimp_pixel_rgn_get_rect(&rgn_in, data, 0, 0, width, height);
-  gimp_drawable_detach(drawable);
-
-  img.image_data = new guchar[pixels * STX_NUM_CHANNELS];
-
-  for (size_t i = 0; i < pixels; i++) {
-    size_t src_pos = i * offset;
-    size_t dst_pos = i * STX_NUM_CHANNELS;
-    // RGBA -> BGRA
-    img.image_data[dst_pos] = data[src_pos + 2];
-    img.image_data[dst_pos + 1] = data[src_pos + 1];
-    img.image_data[dst_pos + 2] = data[src_pos];
-    img.image_data[dst_pos + 3] = has_alpha ? data[src_pos + 3] : 0xff;
-  }
-
-  delete[] data;
-
-  return write(img, file);
 }
